added reports files to reports directory
